update to 0411 version

+++ src/setup.c
@@ -22,6 +22,21 @@
 #include "setup.h"
 
 /*
+** Increment the "cfgcnt" variable, so that ETags will know that
+** the configuration has changed.
+*/
+void setup_incr_cfgcnt(void){
+  static int once = 1;
+  if( once ){
+    once = 0;
+    db_multi_exec("UPDATE config SET value=value+1 WHERE name='cfgcnt'");
+    if( db_changes()==0 ){
+      db_multi_exec("INSERT INTO config(name,value) VALUES('cfgcnt',1)");
+    }
+  }
+}
+
+/*
 ** Output a single entry for a menu generated using an HTML table.
 ** If zLink is not NULL or an empty string, then it is the page that
 ** the menu entry will hyperlink to.  If zLink is NULL or "", then
@@ -166,7 +181,7 @@ void setup_ulist(void){
     style_header("User List");
     @ <table border=1 cellpadding=2 cellspacing=0 class='userTable'>
     @ <thead><tr>
-    @   <th>UID <th>Category
+    @   <th>Category
     @   <th>Capabilities (<a href='%R/setup_ucap_list'>key</a>)
     @   <th>Info <th>Last Change</tr></thead>
     @ <tbody>
@@ -182,7 +197,6 @@ void setup_ulist(void){
       const char *zCap = db_column_text(&s, 2);
       const char *zDate = db_column_text(&s, 4);
       @ <tr>
-      @ <td><a href='setup_uedit?id=%d(uid)'>%d(uid)</a>
       @ <td><a href='setup_uedit?id=%d(uid)'>%h(zLogin)</a>
       @ <td>%h(zCap)
 
@@ -211,12 +225,12 @@ void setup_ulist(void){
   @ </tbody></table>
   @ <div class='section'>Users</div>
   @ <table border=1 cellpadding=2 cellspacing=0 class='userTable sortable' \
-  @  data-column-types='nktxTTK' data-init-sort='2'>
+  @  data-column-types='ktxTTK' data-init-sort='2'>
   @ <thead><tr>
-  @ <th>ID<th>Login Name<th>Caps<th>Info<th>Date<th>Expire<th>Last Login</tr></thead>
+  @ <th>Login Name<th>Caps<th>Info<th>Date<th>Expire<th>Last Login</tr></thead>
   @ <tbody>
   db_multi_exec(
-     "CREATE TEMP TABLE lastAccess(uname TEXT PRIMARY KEY, atime REAL) WITHOUT ROWID;"
+    "CREATE TEMP TABLE lastAccess(uname TEXT PRIMARY KEY, atime REAL) WITHOUT ROWID;"
   );
   if( db_table_exists("repository","accesslog") ){
     db_multi_exec(
@@ -224,7 +238,8 @@ void setup_ulist(void){
       " SELECT uname, max(mtime) FROM ("
       "    SELECT uname, mtime FROM accesslog WHERE success"
       "    UNION ALL"
-      "    SELECT login AS uname, rcvfrom.mtime AS mtime FROM rcvfrom JOIN user USING(uid))"
+      "    SELECT login AS uname, rcvfrom.mtime AS mtime"
+      "      FROM rcvfrom JOIN user USING(uid))"
       " GROUP BY 1;"
     );
   }
@@ -258,7 +273,6 @@ void setup_ulist(void){
       zAge = human_readable_age(rNow - rATime);
     }
     @ <tr>
-    @ <td><a href='setup_uedit?id=%d(uid)'>%d(uid)</a>
     @ <td data-sortkey='%h(zSortKey)'><a href='setup_uedit?id=%d(uid)'>%h(zLogin)</a>
     @ <td>%h(zCap)
     @ <td>%h(zInfo)
@@ -453,7 +467,7 @@ void user_edit(void){
   ** modified user record.  After writing the user record, redirect
   ** to the page that displays a list of users.
   */
-  doWrite = cgi_all("login","info","pw") && !higherUser;
+  doWrite = cgi_all("login","info","pw") && !higherUser && cgi_csrf_safe(1);
   if( doWrite ){
     char c;
     char zCap[50], zNm[4];
@@ -499,6 +513,7 @@ void user_edit(void){
        "VALUES(nullif(%d,0),%Q,%Q,%Q,%Q,now())",
       uid, zLogin, P("info"), zPw, zCap
     );
+    setup_incr_cfgcnt();
     admin_log( "Updated user [%q] with capabilities [%q].",
                zLogin, zCap );
     if( atoi(PD("all","0"))>0 ){
@@ -603,9 +618,10 @@ void user_edit(void){
 
   /* Begin generating the page
   */
-  style_submenu_element("Cancel", cgi_referer("setup_ulist"));
+  style_submenu_element("Cancel", "%s", cgi_referer("setup_ulist"));
   if( uid ){
     style_header("Edit User %h", zLogin);
+    style_submenu_element("Access Log", "%R/access_log?u=%t", zLogin);
   }else{
     style_header("Add A New User");
   }
@@ -1718,7 +1734,7 @@ void setup_adunit(void){
     return;
   }
   db_begin_transaction();
-  if( P("clear")!=0 ){
+  if( P("clear")!=0 && cgi_csrf_safe(1) ){
     db_multi_exec("DELETE FROM config WHERE name GLOB 'adunit*'");
     cgi_replace_parameter("adunit","");
   }
@@ -1807,7 +1823,9 @@ void setup_logo(void){
     return;
   }
   db_begin_transaction();
-  if( P("setlogo")!=0 && zLogoMime && zLogoMime[0] && szLogoImg>0 ){
+  if( !cgi_csrf_safe(1) ){
+    /* Allow no state changes if not safe from CSRF */
+  }else if( P("setlogo")!=0 && zLogoMime && zLogoMime[0] && szLogoImg>0 ){
     Blob img;
     Stmt ins;
     blob_init(&img, aLogoImg, szLogoImg);
@@ -1942,7 +1960,7 @@ int raw_sql_query_authorizer(
 ** Requires Admin privileges.
 */
 void sql_page(void){
-  const char *zQ = P("q");
+  const char *zQ;
   int go = P("go")!=0;
   login_check_credentials();
   if( !g.perm.Setup ){
@@ -1951,6 +1969,7 @@ void sql_page(void){
   }
   add_content_sql_commands(g.db);
   db_begin_transaction();
+  zQ = cgi_csrf_safe(1) ? P("q") : 0;
   style_header("Raw SQL Commands");
   @ <p><b>Caution:</b> There are no restrictions on the SQL that can be
   @ run by this page.  You can do serious and irrepairable damage to the
@@ -2272,6 +2291,7 @@ static void setup_update_url_alias(
   const char *zNewName,
   const char *zValue
 ){
+  if( !cgi_csrf_safe(1) ) return;
   if( zNewName[0]==0 || zValue[0]==0 ){
     if( zOldName[0] ){
       blob_append_sql(pSql,
